move docker resource creation to its own package
