+++ tests/e2e/e2e_test.go
@@ -9,14 +9,14 @@ import (
 // medians deviations are correct, updates the oracle params with
 // a gov prop, then checks the medians and median deviations again.
 func (s *IntegrationTestSuite) TestMedians() {
-	err := grpc.MedianCheck(s.ojoClient)
+	err := grpc.MedianCheck(s.orchestrator.OjoClient)
 	s.Require().NoError(err)
 }
 
 // TestUpdateOracleParams updates the oracle params with a gov prop
 // and then verifies the new params are returned by the params query.
 func (s *IntegrationTestSuite) TestUpdateOracleParams() {
-	params, err := s.ojoClient.QueryClient.QueryParams()
+	params, err := s.orchestrator.OjoClient.QueryClient.QueryParams()
 	s.Require().NoError(err)
 
 	s.Require().Equal(uint64(5), params.HistoricStampPeriod)
@@ -24,12 +24,12 @@ func (s *IntegrationTestSuite) TestUpdateOracleParams() {
 	s.Require().Equal(uint64(20), params.MedianStampPeriod)
 
 	err = grpc.SubmitAndPassProposal(
-		s.ojoClient,
+		s.orchestrator.OjoClient,
 		grpc.OracleParamChanges(10, 2, 20),
 	)
 	s.Require().NoError(err)
 
-	params, err = s.ojoClient.QueryClient.QueryParams()
+	params, err = s.orchestrator.OjoClient.QueryClient.QueryParams()
 	s.Require().NoError(err)
 
 	s.Require().Equal(uint64(10), params.HistoricStampPeriod)
