+++ tests/e2e/orchestrator/chain.go
@@ -1,4 +1,4 @@
-package e2e
+package orchestrator
 
 import (
 	"fmt"
