+++ tests/e2e/orchestrator/oracle_gen_state.go
@@ -1,4 +1,4 @@
-package e2e
+package orchestrator
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
